Change Alloc/Dealloc to Open/Close.
Open returns a Maybe PinDescriptor, where PinDescriptor is (will be) an
abstraction for the resource required by a particular GPIO
implementation for a given GPIO pin; e.g., for Linux /sysfs/gpio this
will be a file descriptor.

Using a descriptor allows us to require that a pin has been
initialized (via Open) before being used.