Only Open returns Either; others should throw.
This makes it easier to write clean GpioT computations and to move the
error-handling elsewhere. The beauty of the Free monad/interpreter
approach is that we can implement exceptions differently depending on
the interpreter. For example, the runMock interpreter does it via
ExceptT String so that all exceptions are caught with a Left String.