Add 'availablePins' to DSL.
Makes a best-effort attempt to list all the GPIO pins available on the
system. Note that on some platforms, this may return an empty list (or a
partial list of what's actually available). For the Mock interpreter, it
returns the complete list. For the Linux sysfs interpreter, it returns
whatever GPIO chips are exposed by the /sys/class/gpio interface (so,
basically, everything that's been exported by the kernel via the sysfs
interface).