First-pass attempt at a 'MockSysfs' monad.
Unlike most of the other types in the gpio package, 'MockSysfs' is
neither an interpreter nor a Free monad. Instead, it mimics all of the
filesystem operations performed by the 'SysfsT' interpreter/transformer,
so that you can use 'runSysfsT' and friends on platforms that don't
support the Linux sysfs GPIO interface, or in situations (e.g., testing)
where you want to run GPIO programs that don't need root access (which
is typically required to manipulate sysfs GPIO state on most Linux
systems).

Though it hasn't been integrated into the test framework yet (that will
come in a subsequent check-in), the idea is that 'MockSysfs' can be
dropped into the 'SysfsT' transformer stack and thereby replace the
actual sysfs I/O operations with its own mock operations, so that no
actual Linux sysfs GPIO implementation is needed to test the 'SysfsT'
program with a reasonable degree of certainty about its proper
operation (to the extent that 'MockSysfs' faithfully emulates the real
Linux sysfs GPIO interface, at least).

I believe this is a better approach to mocking/testing than the existing
one (the 'MockT' transformer), which implements a mock interpreter with
no underlying real-world GPIO implementation. In fact, once the
'MockSysfs' monad is integrated into the gpio package test framework, I
plan to remove the 'MockT' interpreter.

Among the reasons why I believe it's a good idea to remove 'MockT' and
replace it with 'MockSysfs':

* Prior to this check-in, there was no good way to test the 'SysfsT'
  interpreter. Currently, the gpio package test framework uses the
  'MockT' interpreter, which doesn't really test anything particularly
  useful, since the interpreter itself is never used for real-world
  programs, and because the test programs themselves are also not very
  interesting or useful outside the testing environment. About the only
  useful thing the 'MockT' interpreter provided was a way to "sound out"
  the development of the 'GpioF' DSL, in order to ensure that the
  language was expressive enough to make useful GPIO programs. Now that
  the language has reached ~"1.0" status, 'MockT' has outlived its
  usefulness.

  Once the test harnesses are re-written to use 'MockSysfs' with the
  'SysfsT' transformer, rather than 'MockT', the tests will be testing a
  real-world interpreter. Assuming that 'MockSysfs' does a reasonable
  job of emulating the actual Linux sysfs API, we should be able to
  catch most bugs in the 'SysfsT' transformer/interpreter in the test
  environment, even on development platforms with no actual GPIO
  support (e.g., OS X).

* 'MockSysfs' could also be used for prototyping and testing other
  programs which use the gpio package. For example, my "mellon-gpio"
  program (part of the "mellon" package) uses the 'SysfsT' transformer
  to control physical access devices. Prior to the invention of
  'MockSysfs', testing mellon-gpio required testing against actual
  access devices. This was inconvenient, and effectively required that
  development related to testing be performed on the same low-power
  embedded devices on which the production system is hosted (usually one
  and the same).

  Once mellon-gpio is updated to use the new version of gpio with the
  'MockSysfs' monad, however, it should be possible to develop and test
  mellon-gpio on any system, with a reasonable expectation that, once
  the program is believed to be robust on the development system, there
  should be much less time and effort needed to vet it on the actual
  production hardware.

* As 'MockSysfs' obviates the need for an entire faux-GPIO interpreter,
  there's no need to maintain 'MockT' beyond this point. 'MockSysfs' is
  probably roughly the same complexity as the 'MockT' interpreter,
  mainly because of the lengths to which it goes to emulate the
  exceptions that can be thrown by the real Linux sysfs filesystem; but
  of course it comes with a host of other benefits, as described above.

I probably should have started with this approach from the beginning,
but it didn't occur to me until too late in the development cycle. This
has been a good lesson and I think I'll take this approach in future
projects. At the very least, when I implement new GPIO interpreters in
this package (e.g., for OpenBSD), I'll certainly develop mock system
emulators for testing those interpreters.