'IO' can't be an instance of 'MonadSysfs', move to a new typeclass.
The new typeclass is called 'SysfsIOT' and captures the "real" Linux
sysfs GPIO implementation in a monad transformer. It comes with its own
"run" functions which, when used, will run a 'SysfsF'/'GpioF' program on
the actual Linux sysfs filesystem.

The 'IO' functionality had to be moved out of 'MonadSysfs' because the
'IO' instance of 'MonadSysfs' was overriding all other 'MonadSysfs'
implementations in a given stack, due to the fact that the 'IO' monad
must always be the base of any stack in which it appears. Because
'MockSysfsT' (now renamed 'SysfsMockT' to match the 'SysfsIOT' naming
convention) was also an instance of 'MonadIO', any stack in which it
appeared also had an 'IO' at the bottom. Therefore, the 'MockSysfsT'
mock implementation of the Linux sysfs interface was always overridden
by the 'IO' instance of 'MonadSysfs'; i.e., there was no way to use the
mock implementation!

By moving the 'IO' implementation of 'MonadSysfs' to a separate
'SysfsIOT' transformer, the user now has to make an explicit choice to
use either 'runSysfsMockT' or 'runSysfsIOT' to run the computation, and
this ensures that the proper 'MonadSysfs' implementation will be used
while running that computation.