Add convenience run* functions for 'SysfsMockT'.
A note on 'runSysfsMockSafe': though it converts 'IOError's that occur
as a side effect of buggy 'SysfsT' programs to 'Left' 'String' for
exception safety, it cannot capture the 'MockWorld' state and log output
under these conditions, because the 'SysfsMockT' layer raises 'IOError's
which don't contain that information, and the raised errors escape the
'SysfsMockT' monad in the stack -- they must, because 'scriptIO', which
we use to catch these errors and convert them to 'Left' 'String', works
at the 'IO' layer.

It should be possible to modify 'SysfsMockT' to provide the state and
log when raising these errors, so that the error type itself carries the
info, but it may have knock-on effects on the type signature of
'SysfsT', so this will require some investigation. It may also mean
changing the error type from 'IOError' to something else, which means
that the fidelity of 'SysfsMockT's 'IO' emulation would be worsened.