Make MonadSysfs more type-safe.
Previously the readPin* functions in the MonadSysfs typeclass returned
String, where the string value was the raw value read from the pin's
sysfs "attribute" files. That was fine when the primary user interface
was meant to be the GpioF eDSL. However, now that I'm promoting
MonadSysfs to first-class status along with the eDSL, it makes sense for
these functions to return something more type-safe, so now SysfsIO does
the raw IO String -> X translation, rather than punting it to the SysfsT
interpreter.