Add code to EdgeRead to test readPin with timeout.
It works! It's a bit awkward due to the fact that timeout wants an IO
action, so you have to run the interpreter again within that action to
make the types work. This, in turn, means that the type of the
subprogram that runs 'readPin' can't be GpioT h m m; you need to choose
a particular (IO-based) interpreter type.

This *might* be a bit more composable if the "pin handle" type 'h' in
the GpioT h m m signature were concrete (i.e., the same type for all
interpreters); then the computation which opens the pin and obtains the
handle could lift the handle into the IO action and pass it to the
interpreter. However, that seems like a bad idea as it would limit the
flexibility that interpreters have to choose their own handle type. For
example, for Linux sysfs you don't really need anything other than the
original Pin value, as a "pin" in sysfs is broken into constituent files
in sysfs; whereas, for an ioctl-based system (e.g., FreeBSD), you
probably want a standard Haskell resource handle to represent the pin
handle.

Another option would be to abstract 'timeout' and build it into the
DSL (or at least into the readPin command). This might be a better
approach as it works well with the poll(2) mechanism used in the Linux
sysfs 'readPin' implementation, and would presumably work well with
ioctl-based APIs, as well.