Add readPinTimeout to eDSL.
I decided to implement this directly in the eDSL, rather than requiring
the user to use Haskell's own 'timeout' function. The latter doesn't
compose well with the eDSL since the eDSL is its own monad, and
'timeout' requires its action to be IO a. With this change, the user can
specify a timeout on the readPin command directly in the DSL, without
needing to go all the way back to IO. This greatly simplifies programs
which try to do this; see the changes to the Gpio.hs example program for
proof. It also provides some flexibility & room for optimization in
terms of the implementation, at the platform/interpreter level.

Speaking of which -- the current implementation of this in SysfsIO is a
bit limited in that it uses poll(2)'s own timeout functionality to
implement the timeout. poll(2) takes a timeout specified in
milliseconds, whereas Haskell generally uses microsecond-precision
timeouts/delays. One alternative to the current implementation which
might provide better precision is to implement the poll(2) wrapper as
before -- using -1 for the timeout so that it blocks indefinitely -- and
to use Haskell's 'timeout' function to kill the blocked C FFI call when
the timeout expires. However, the poll(2) wrapper is *already* wrapped
with 'throwErrnoIfMinus1Retry' in order to implement proper error- and
exception-/throwTo-handling from Haskell threads. Adding another layer
of complex Haskell thread magic on top of 'throwErrnoIfMinus1Retry' and
the C FFI seems like a bad idea, not to mention the fact that it *must*
be less efficient than taking advantage of poll(2)'s own built-in
timeout feature!

As this code is intended to run on Linux, I think the proper fix is to
use Linux's ppoll(2), which has a higher-precision timeout argument.
However, as that isn't available on other platforms (notably the BSDs
and OS X), I won't be able to compile code which uses it on those
platforms. From a runtime functionality perspective, this doesn't
matter, as none of the SysfsIO code will do anything useful on those
platforms, anyway. However, it is convenient for me to ensure that
SysfsIO.hs contains syntactically correct Haskell code on OS X, as I
develop on that platform, so I'm unwilling to #ifdef away the entire
contents of the module on non-Linux platforms. What I'll probably end up
doing is #ifdef-ing only around the poll(2)/ppoll(2) call, but I need to
figure out how that interacts with the inline-c package that I'm using.
Anyway, that's left for a future improvement; this version should work,
if only with lesser timeout precision.

p.s. another change in this commit is a bug fix in Gpio.hs that was
leftover from a previous change, where the timeout/delay units were
inconsistent.