Simplify Mock filesystem types.
Eliminate MockFS type, replace it with a Rose tree-like structure where
the list of sub-trees is always directories, not directories and files.
This eliminates a whole class of error results because it's no longer
possible to "goto" a file. The list of files in a directory is now
treated as a separate list and is not part of the tree structure, but
instead just an element of each node.

Note that the 'cd' command here cannot handle absolute paths (i.e.,
paths that begin with "/"). That will be fixed in a subsequent commit.