Change exception value when 'cd' fails.
Strictly speaking, 'cd' should return 'NotADirectory' when a name
names a file in that directory, and 'NoSuchFileOrDirectory' when name
doesn't exist, but it always returns 'NotADirectory' to keep things
simple. (This is a consequence of the simpler design that treats files
and directories differently.)

This could be rather easily changed in a subsequent update if so
desired.