Switch from 'Free' to mtl-style type classes.
After much deliberation, I've decided not to use free monads (nor any of
its variants, e.g., "extensible effects") to implement the interpreters
in this package. I was persuaded mainly by the performance arguments:
type classes are simply much more efficient (by at least a factor of
two) than any free monad/extensible effects implementation at the time
of writing.

Because most of the interpreters in the package aren't implemented on
top of any existing mtl class, they don't suffer from what I perceive to
be the primary drawback of mtl-style interpreters/type classes --
namely, the fact that using multiple instances of an mtl type class in a
given transformer stack is unwieldy/impossible (depending on whether
you're willing to turn on certain GHC extensions). Because most of the
interpreters here are stand-alone, application writers who use this
package can freely combine these interpreters with their own transformer
stacks and should not have to worry about conflicts or compositional
problems.

It appears to me that the chief complaints about mtl-style type class
interpreters are a) the amount of boilerplate required to make 'lift'ing
possible across the various mtl types, and b) a general unease about
defining type classes which lack laws. Regarding a), this is simply an
implementation issue which only affects the library writer, not the
library's users, and I am willing to do the manual lifting if it means
much better performance.

Regarding b), while I don't disagree with the feeling that it's a good
thing when type classes specify laws in order to provide some guidance
to instance writers about how a certain type class method should
behave (e.g., 'MonadReader' 'ask' should not modify its environment), I
think that in the case of interpreter-like type classes, because the
interpreters are generally so specialized to their problem
domains (unlike the universally useful mtl type classes, e.g.,
MonadReader), lack of laws is mostly a non-issue. For one thing, it's
quite common for interpreters to perform operations in some subset of
'IO' (or mock 'IO'), where it's difficult to provide /any/ guarantees
about the behavior of interpreter actions. It's not clear to me that
laws provide any real value for these types of interpreters, if any laws
can be identified for them in the first place.

Furthermore, by using a lawless type class to define the grammar of our
interpreters, it's not like we're /losing/ any laws relative to a free
monad or (especially) an extensible effects approach. The whole point of
those approaches is that they make it very easy to provide "types a la
carte" and eliminate many of the rigorous aspects of providing effects.
Indeed, other than performance objections, it seems that the community's
next-most common complaint about a free monad/extensible effects
approach to effects is that they do not provide any guarantees about
which other (often mtl-provided) effects they're compatible with. In
other words, by using lawless type classes rather than free monads or
extensible effects to implement our interpreters, we're not any worse
off with respect to laws.

In fact, at least by specifying mtl-style type class instances of our
interpreters, at least we can easily provide a list (via
GeneralizedNewtypeDeriving and a suitable 'deriving' clause) of other
mtl type classes which /are/ compatible with our interpreter
implementation. (The same can be said of the free monad approach, of
course.)

So: to me, other than the extra boilerplate required, there appear to be
no drawbacks to this type class approach vs. the previous free monad
approach; while there is, at least, the benefit of a fairly significant
performance improvement.