sysfs 'pinHasDirection' and 'pinHasEdge' can throw.
These actions previously did not throw, but returned 'Nothing' when the
pin being queried was not exported. Indicating that a pin doesn't
support these actions should be handled differently than trying to query
a pin that hasn't been exported (the latter is clearly a programming
error), so add a new 'NotExported' exception value to 'SysfsException'
and use it when appropriate.