Change the way the sysfs direction attribute is handled.
Bizarrely, in Linux sysfs GPIO, when you write a pin's value attribute,
the pin's active-low setting is taken into account; but when you write
the pin's direction attribute with "high" or "low", the pin's active-low
setting is /not/ taken into account. In other words, when you write the
pin's direction attribute with a signal level (rather than just "in" or
"out), you are setting the pin's /physical/ signal level, but when you
write its value attribute, you are setting its /logical/ signal level.

I find this very surprising (and confusing), so I've changed the
behavior of the Linux sysfs GPIO interpreter to treat writes to the
direction attribute as a logical value.

This change also makes a bunch of documentation updates to be more
consistent about the pin's "signal level", logical vs. physical values,
etc.