Graceful handling of double-open and double-close.
On Linux it's an error to export a pin that's already exported, or to
unexport a pin that's not already exported. This isn't particularly
useful behavior so in the portable DSL we just ignore these errors.

The low-level Linux DSL does provide actions which enforce these errors,
if you need them.