First pass at a "safer" cross-platform DSL.
Not really a DSL per se, but actions and types built on top of the
existing cross-platform DSL which make it impossible to perform illegal
actions on pins in certain pin configurations. Namely:

- An 'InputPin' type that can only be read.
- An 'OutputPin' type that can be read or written.
- An 'InterruptPin' type that can be read or polled.

If you stick to the safer DSL, you don't have to worry about all of the
odd corner cases that can bite you, like trying to write a pin that's
been configured for input, or setting the interrupt mode on an output
pin.

Furthermore, the constructors for the safer types are not exported; they
can only be created by 'with*' wrappers which will fail if the pin
cannot be configured in the requested fashion. For example,
'withInterruptPin' only succeeds if the pin supports interrupts.

There is one bit of ugliness: whereas previously we only had concrete
exception types at the platform interpreter level, we have now added an
exception type at the DSL layer. Currently, the only use case for this
new exception type is to be able to convert an 'InterruptPin''s
interrupt mode from 'Maybe' 'PinInterruptMode' to plain
'PinInterruptMode' and handle the ("impossible") case where a pin that
was successfully configured for interrupts (as an 'InterruptPin')
suddenly no longer supports interrupts when its interrupt mode is
subsequently queried.

I don't particularly like this design choice, but if this scenario ever
actually occurs, it truly is exceptional! And it seems wrong to leave
the value wrapped in a 'Maybe' when the fact that you were able to
construct the 'InterruptPin' in the first place effectively guarantees
that the value should never be 'Nothing'.