Remove input/output mode settings for safer pin types.
These were another source of possible run-time configuration
errors (e.g., you could create an 'InputPin' using a supported mode,
then try to change its mode to something unsupported and raise an
exception).

With this change, I believe it shouldn't be possible for the programmer
to cause an exception on an 'InputPin, 'InterruptPin', or 'OutputPin'
type.