This project now uses Protolude.
As part of this change, I've gotten rid of String everywhere except a)
where required by a dependency, e.g., for FilePath; or b) where it is
necessary to use for some low-level system API or emulation of that
API (e.g., one or two places in the Mock Linux sysfs interpreter).

Note: Tutorial.hs still uses Prelude rather than Protolude, as I
didn't want to throw a curveball at readers who are unfamiliar with
Protolude.

Note: this commit will definitely require a PVP bump, as some of these
changes are visible in the public API.