By default, pass an empty 'overlays' to nixpkgs.
I noticed that I was suddenly unable to "cabal test", with an error
from cabal similar to that seen in
https://github.com/haskell/cabal/issues/1665. I could not figure out
the problem until I tried to run 'nix-build release.nix' and noticed
that nix-build was not running the unit tests at all.

I traced this back to a 'dontCheck' on the upstream nixpkgs hpio
package in my Mac's nixpkgs overlays. The overlay was being applied to
the local release.nix override of hpio, so it was not being checked,
and when I asked cabal to run the tests explicitly, the 'dontCheck'
caused this cryptic error.

The lesson is to override overlays in release.nix, because in most
cases we want to build a reproducible binary no matter what the user's
local environment (modulo their version of nixpkgs, of course -- I
will address this in a subsequent commit). However, we do allow the
user to override the overlays, if needed, by passing them as an
argument to release.nix.