Add nixpkgs to jobset inputs.
nixpkgs-stackage imports from <nixpkgs> to get Haskell functions, so
we need to make sure it's set.

Because nixpkgs-stackage is overlay-based, it gets its packages from
the nixpkgs passed to the overlay, so the nixpkgs used for <nixpkgs>
isn't particularly important. For the default job (hpio master
branch), we just use the nixpkgs passed to nix/jobset/default.nix. For
the alt jobs, we use the value of nixpkgs_override.