More Hydra job cleanup.
Not much point in enumerating aggregate job constituents if they all
have the same name. (It did not work the way I'd hoped.) Therefore,
don't bother with that and just let Hydra list the aggregate jobs.

Also, move the "simple" hpio-on-the-current-platform Makefile target
to testing.nix so that it doesn't show up in the Hydra results.